refactor: column, row 객체 구현
